requirements.txt
Flask
qrcode
pycryptodome
gunicorn
Pillow
